api/entities: refactor entities to use an index file for cleaner usage in imports
