+++ core_api/src/database/dataSource.ts
@@ -1,14 +1,16 @@
 import * as dotenv from 'dotenv';
 import { DataSource } from 'typeorm';
-import { Role } from '../modules/role/role.entity';
-import { Department } from '../modules/department/department.entity';
-import { Gender } from '../modules/gender/gender.entity';
-import { User } from '../modules/user/user.entity';
-import { WorkingHours } from '../modules/working_hours/workingHours.entity';
-import { Attachment } from '../modules/attachment/attachment.entity';
-import { Consultation } from '../modules/consultation/consultation.entity';
-import { ConsultationSubject } from '../modules/consultation_subject/consultationSubject.entity';
-import { Patient } from '../modules/patient/patient.entity';
+import {
+  Attachment,
+  Consultation,
+  ConsultationSubject,
+  Department,
+  Gender,
+  Patient,
+  Role,
+  User,
+  WorkingHours
+} from '../modules/entities.index';
 
 dotenv.config();
 const { NODE_ENV } = process.env;
