+++ core_api/src/modules/consultation_subject/consultationSubject.entity.ts
@@ -6,7 +6,7 @@ import {
   OneToMany
 } from 'typeorm';
 import { Field, ObjectType, Int } from 'type-graphql';
-import { Consultation } from '../consultation/consultation.entity';
+import { Consultation } from '../entities.index';
 
 @ObjectType()
 @Entity()
