+++ core_api/src/modules/entities.index.ts
@@ -1,0 +1,9 @@
+export * from './attachment/attachment.entity';
+export * from './consultation/consultation.entity';
+export * from './consultation_subject/consultationSubject.entity';
+export * from './department/department.entity';
+export * from './gender/gender.entity';
+export * from './patient/patient.entity';
+export * from './role/role.entity';
+export * from './user/user.entity';
+export * from './working_hours/workingHours.entity';
