+++ core_api/src/modules/role/role.entity.ts
@@ -6,7 +6,7 @@ import {
   OneToMany
 } from 'typeorm';
 import { Field, ObjectType, Int } from 'type-graphql';
-import { User } from '../user/user.entity';
+import { User } from '../entities.index';
 
 @ObjectType()
 @Entity()
